Move StressTestClient
